+++ src/main/java/com/aurd/Student/Model/Entity/Current_AffairsCommented_Model.java
@@ -31,7 +31,7 @@ public class Current_AffairsCommented_Model {
     private Timestamp added_on;
 
     @Column(name = "added_by",nullable = false)
-    private int added_by;
+    private Long added_by;
 
     @Column(name = "type",nullable = false)
     private String type;
@@ -94,11 +94,11 @@ public class Current_AffairsCommented_Model {
         this.current_affair_id = current_affair_id;
     }
 
-    public int getAdded_by() {
+    public Long getAdded_by() {
         return added_by;
     }
 
-    public void setAdded_by(int added_by) {
+    public void setAdded_by(Long added_by) {
         this.added_by = added_by;
     }
